move 'SyncStorage' to /tests.
