+++ src/storage/mod.rs
@@ -25,22 +25,17 @@ pub mod mvcc;
 pub mod txn;
 pub mod config;
 mod types;
-mod sync_storage;
 
 pub use self::config::Config;
 pub use self::engine::{Engine, Snapshot, Dsn, TEMP_DIR, new_engine, Modify, Cursor,
                        Error as EngineError};
 pub use self::engine::raftkv::RaftKv;
 pub use self::txn::{SnapshotStore, Scheduler, Msg, SchedCh};
-pub use self::types::{Key, Value, KvPair};
-pub use self::sync_storage::SyncStorage;
+pub use self::types::{Key, Value, KvPair, make_key};
 pub type Callback<T> = Box<FnBox(Result<T>) + Send>;
 
 pub type CfName = &'static str;
 pub const DEFAULT_CFS: &'static [CfName] = &["default", "lock"];
-
-#[cfg(test)]
-pub use self::types::make_key;
 
 #[derive(Debug, Clone)]
 pub enum Mutation {
