+++ tests/coprocessor/test_select.rs
@@ -4,7 +4,7 @@ use kvproto::kvrpcpb::Context;
 use tikv::util::codec::{table, Datum, datum};
 use tikv::util::codec::datum::DatumDecoder;
 use tikv::util::codec::number::*;
-use tikv::storage::{Dsn, Mutation, Key, SyncStorage, DEFAULT_CFS};
+use tikv::storage::{Dsn, Mutation, Key, DEFAULT_CFS};
 use tikv::storage::engine::{self, Engine, TEMP_DIR};
 use tikv::util::event::Event;
 use tikv::util::worker::Worker;
@@ -12,6 +12,7 @@ use kvproto::coprocessor::{Request, KeyRange};
 use tipb::select::{ByItem, SelectRequest, SelectResponse};
 use tipb::schema::{self, ColumnInfo};
 use tipb::expression::{Expr, ExprType};
+use storage::sync_storage::SyncStorage;
 
 use std::sync::Arc;
 use std::collections::{HashMap, BTreeMap};
